requirements.txt
fastapi
uvicorn
mistralai
pydantic
python-dotenv
